% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_value.R
\name{p_value}
\alias{p_value}
\title{P_Value of a model performance metric}
\usage{
p_value(true_score, shuffled_scores, report = FALSE)
}
\arguments{
\item{true_score}{numeric metric score on non-shuffled data}

\item{shuffled_scores}{numeric vector of metrics on shuffled data.
Will have length equal to the number of simulations used during
target shuffling.}
}
\value{
numeric proportion of shuffled scores that are lower than true score.
  This can be interpreted as a p_value measuring the significance
  of the model's performance.
}
\description{
P_Value of a model performance metric
}
\examples{
p_value(true_precision, shuffled_precisions)
}
